feat: Add summer theme to kiosk
This commit adds a new theme option, "summer", to the kiosk application. When the "summer" theme is selected in the admin.html file, a wave animation is displayed at the bottom of the kiosk page. This feature enhances the visual experience for users during the summer season.

Recent user commits:
- fix transprent images
- fix animation :D
- shit transition ill never fix
- Update kiosk.css
- a
- Update kiosk.css
- fixed couldnt edit
- feat: add Loading spinner in kiosk
- feat: Enable save button when editing kiosk item
- Revert "Update script.js"