feat: Add Halloween theme to kiosk
This commit adds a new theme option, "Halloween", to the kiosk application. When the "Halloween" theme is selected in the admin.html file, a snowflake animation with Halloween-themed emojis is displayed at the bottom of the kiosk page. This feature enhances the visual experience for users during the Halloween season.